General site wide tweaks
Made the logo smaller, adjusted the padding between show categories on home page and shows page. Reduced number of shows per line and truncated show title. Adjusted look of the pagination. Adjusted headings to be uniform. Added borders to video thumbnails so the white isn't lost in the background. Added padding to the metadata area on single show page. Adjust site size to be wider.