Main site area tweaks
Adjusted the show title lengths, made thumbnails bigger and less per line. Adde borders to different sections, moved the nagivation links further apart. Moved the News and Partners sections into the template code instead of on the page itself and made them only visible if there is content. Adjusted the text on the news articles pages and the show detail pages to read nicer and be more consistent with the rest of the site.