Finalizing design tweaks 6 5 (#7)
* General site wide tweaks

Made the logo smaller, adjusted the padding between show categories on home page and shows page. Reduced number of shows per line and truncated show title. Adjusted look of the pagination. Adjusted headings to be uniform. Added borders to video thumbnails so the white isn't lost in the background. Added padding to the metadata area on single show page. Adjust site size to be wider.

* Main site area tweaks

Adjusted the show title lengths, made thumbnails bigger and less per line. Adde borders to different sections, moved the nagivation links further apart. Moved the News and Partners sections into the template code instead of on the page itself and made them only visible if there is content. Adjusted the text on the news articles pages and the show detail pages to read nicer and be more consistent with the rest of the site.

---------

Co-authored-by: Teresa Loafman <66372353+teresasanders77@users.noreply.github.com>